Create a game now works
The create a game button now works.
OBS: If u want to clear the game array in database.json, calla a delete request with your browser.
TODO: Maybe switch the games object to a STATE? FIx so that the join button works and other "users" (connections) can connect to created games